Move utilCustomEvents to base/events
Caleydo/cohort#693
