+++ src/Taskview/columns/AColumn.ts
@@ -3,7 +3,7 @@ import { Cohort, EMPTY_COHORT_ID, LOADER_COHORT_ID } from '../../Cohort';
 import { colors } from '../../config/colors';
 import { createSearchBarTooltip } from '../../Tooltip';
 import { getAnimatedLoadingBars, log } from '../../util';
-import { ColumnCloseEvent } from '../../utilCustomEvents';
+import { ColumnCloseEvent } from '../../base/events';
 import Taskview, { InputCohort, OutputCohort } from '../Taskview';
 
 export abstract class AColumn {
