+++ src/Taskview/visualizations/AVegaVisualization.ts
@@ -9,8 +9,7 @@ import { ICohort } from '../../CohortInterfaces';
 import { IAttribute, IdValuePair } from '../../data/Attribute';
 import { IEqualsList, INumRange, NumRangeOperators } from '../../base/rest';
 import { IFilterDesc, log } from '../../util';
-import { CoralColorSchema } from '../../config/colors';
-import { FilterEvent, SplitEvent } from '../../utilCustomEvents';
+import { FilterEvent, SplitEvent } from '../../base/events';
 import { Option, VisConfig } from './config/VisConfig';
 import { DATA_LABEL } from './constants';
 import { IVisualization } from './IVisualization';
